Complete IpNetwork and IpInterface tests.
This is now sufficient to use to test IpNetwork and IpInterface for integration
tests.

The IpNetwork test allows you to specify an HTTP url, which will be fetched and
its contents placed in a text node on the page.

The IpInterface test allows you to listen on a particular port.  The first connection
to that port will have the string "tcptest" written to it, then the connection will
be closed.

Due to a bug in pycapnp [1], IpInterface continues listening until the grain is shut down.
This does not prevent the test from testing what it was intended to test, though.

[1] - https://github.com/jparyani/pycapnp/issues/90