Rework IpInterface test into a two-RPC flow
Our current IpInterface test is flaky due to a race condition between the
backend listening and the test trying to connect to the opened socket.

The solution is to remove the race, which requires splitting the API into
separate "start listening" and "interact" requests, where we get explicit
confirmation of the listen()'s completion, and only then attempt to connect
to the listening socket.

This change, combined with a change to the tests, should make the IpInterface
test no longer fail spuriously.